fix: use correct stats loading status in token detail
